RNGSystem: Improve precalculation
This now also stores the first 2 criteria for each game which adds an additional byte per seed stored in the file.  This allows to completely skip those in the first pass whose first 2 criteria doesn't match.  Although the performance gain is negligible in xd, in colo, this is a 66% improvements.