RNGSystem: Redo precalculation entirely
This is a massive change, but essentially, it was found out that out of the 4 billions possible seeds, only a very small fraction (less than 1/20) of them were unique after the first team generation.  This meant that it would be possible to recalculate them in a file and already have the correct seeds to work with. Thanks to a suggestion by Thurler, it was possible to improve it further by storing the seeds in the order of the first 2 criteria (flatten the criteria to one index) which also required to store the sizes of the vectors at the header of the file.

Result: this completely removes the concept of seed range in the backend and makes the first pass instant.  However, the GUI will have to be refactored and reimplemented to have the seed finder work.