RNGSystem/XD: Fix the EVs generation function
This is one of the nastiest bug I ever had.

The sum should never be reset at the end of the loop, but at the beginning.  This changes the very rare cases where the last iteration would reset the sum, but not the EVs while the game doesn't reset anything in that case. This missmatch caused the HP to be miscalculated so when it was compared against what the user entered, it would have been incorrect. Therefore, this does not affect the precalc files.